+++ packages/article/data-helper.js
@@ -1,5 +1,21 @@
 import getLeadAsset from "./get-lead-asset";
 
+const prepend = ({ type, data }, list) => {
+  if (!data) {
+    return list;
+  }
+
+  return [{ type, data }, ...list];
+};
+
+const append = ({ type, data }, list) => {
+  if (!data) {
+    return list;
+  }
+
+  return [...list, { type, data }];
+};
+
 const prepareDataForListView = articleData => {
   const { isVideo, leadAsset } = getLeadAsset(articleData);
   const articleHeaderData = {
@@ -14,16 +30,20 @@ const prepareDataForListView = articleData => {
     publishedTime: articleData.publishedTime,
     byline: articleData.byline
   };
+  const relatedArticlesData = articleData.relatedArticles
+    ? {
+        relatedArticles: articleData.relatedArticles,
+        relatedArticlesLayout: articleData.relatedArticlesLayout
+      }
+    : null;
 
   const data = [
-    { type: "leadAsset", data: leadAsset },
     { type: "header", data: articleHeaderData },
-    { type: "middleContainer", data: articleMidContainerData }
-  ].concat(
-    articleData.content.map((rowData, index) => {
+    { type: "middleContainer", data: articleMidContainerData },
+    ...articleData.content.map((rowData, index) => {
       const item = {
         type: "articleBodyRow",
-        data: rowData,
+        data: Object.assign({}, rowData),
         index
       };
       if (rowData.name === "ad") {
@@ -34,13 +54,12 @@ const prepareDataForListView = articleData => {
       }
       return item;
     })
+  ];
+
+  return prepend(
+    { type: "leadAsset", data: leadAsset },
+    append({ type: "relatedArticles", data: relatedArticlesData }, data)
   );
-
-  if (!leadAsset) {
-    data.splice(0, 1);
-  }
-
-  return data;
 };
 
 export default prepareDataForListView;