+++ packages/author-profile/author-profile-error.web.js
@@ -3,7 +3,7 @@ import { Text, View, Button } from "react-native";
 import PropTypes from "prop-types";
 import Image from "@times-components/image";
 import withResponsiveStyles from "@times-components/responsive-styles";
-import { colours } from "@times-components/styleguide";
+import { colours, fonts, fontSizes } from "@times-components/styleguide";
 
 const ErrorContainer = withResponsiveStyles(View, {
   base: () => `
@@ -25,8 +25,8 @@ const ErrorContainer = withResponsiveStyles(View, {
 
 const Heading = withResponsiveStyles(Text, {
   base: () => `
-    font-family: TimesModern-Bold;
-    font-size: 35px;
+    font-family: ${fonts.headline};
+    font-size: ${fontSizes.leadHeadline}px;
     text-align: center;
     color: ${colours.functional.brandColour};
     margin-bottom: 12px;
@@ -38,8 +38,8 @@ const Heading = withResponsiveStyles(Text, {
 
 const Message = withResponsiveStyles(Text, {
   base: () => `
-    font-family: TimesDigitalW04-Regular;
-    font-size: 18px;
+    font-family: ${fonts.bodyRegular};
+    font-size: ${fontSizes.body}px;
     line-height: 1.44;
     text-align: center;
     color: ${colours.functional.secondary};