+++ packages/date-publication/date-publication.stories.js
@@ -1,13 +1,13 @@
 import React from "react";
 import { Text, StyleSheet } from "react-native";
 import { storiesOf } from "@storybook/react-native";
-import { colours, fonts } from "@times-components/styleguide";
+import { colours, fonts, fontSizes } from "@times-components/styleguide";
 import DatePublication from "./date-publication";
 
 const styles = StyleSheet.create({
   text: {
     color: colours.functional.secondary,
-    fontSize: 13,
+    fontSize: fontSizes.cardMeta,
     lineHeight: 15,
     fontFamily: fonts.supporting
   }