+++ packages/pagination/pagination-icons.js
@@ -1,14 +1,14 @@
 import React from "react";
 import { Text, View, StyleSheet, Platform } from "react-native";
 import Svg, { G, Path } from "svgs";
-import { colours, fonts } from "@times-components/styleguide";
+import { colours, fonts, fontSizes } from "@times-components/styleguide";
 
 import PageLabel from "./page-label";
 
 const textStyle = {
   height: 15,
   fontFamily: fonts.supporting,
-  fontSize: 15,
+  fontSize: fontSizes.pagingMeta,
   color: colours.functional.action
 };