+++ packages/slice/__tests__/web/jest.config.js
@@ -1,6 +1,3 @@
 const jestConfigurator = require("@times-components/jest-configurator").default;
 
-module.exports = jestConfigurator("web", __dirname, [
-  "**/templates/**/index.js",
-  "**/templates/styles.js"
-]);
+module.exports = jestConfigurator("web", __dirname, ["index.js"]);