+++ packages/slice/templates/standard/index.web.js
@@ -1,6 +1,6 @@
 import React from "react";
 import propTypes from "./proptypes";
-import { getSeparator, SliceContainer } from "../shared.responsive";
+import { getSeparator, SliceContainer } from "../styles/responsive";
 import { getChildrenContainer, ChildContainer } from "./responsive";
 import config, { getConfigWrapper } from "./config";
 
@@ -16,7 +16,9 @@ const StandardSlice = ({ itemCount, renderItems }) => {
       <SliceContainer>
         <ChildrenContainer>
           {renderItems(config)
-            .map(item => <ChildContainer key={item.key}>{item}</ChildContainer>)
+            .map(item => (
+              <ChildContainer key={item.props.id}>{item}</ChildContainer>
+            ))
             .reduce((previous, current) => [
               ...(previous.length > 0 ? previous : [previous]),
               <Separator key={`separator-${current.key}`} />,